requirements.txt
Click
docker-compose>=1.27.4<=1.29.2
pyyaml>=5.4.1
MarkupSafe
jinja2
